# Performance improvement for queries having join between TABLE_CONSTRAINTS and KEY_COLUMN_USAGE views